requirements.txt
Django===3.1.7
djangorestframework===3.12.3
psycopg2===2.8.6
Pillow>=5.3.0,<5.4.0
django-cors-headers===3.7.0